inv: Add option for 'types' task to create SoftFab results file

+++ tasks.py
@@ -55,23 +55,47 @@ def lint(c, src=None, rule=None):
         c.run('pylint %s' % ' '.join(args), env=PYLINT_ENV, pty=True)
 
 @task
-def types(c, src=None, clean=False, report=False):
+def types(c, src=None, clean=False, report=False, results=None):
     """Check sources with mypy."""
     if clean:
         print('Clearing mypy cache...')
         remove_dir(TOP_DIR / '.mypy_cache')
     print('Checking sources with mypy...')
+    report_dir = None if results is None else Path(results).parent.resolve()
     args = []
     if report:
-        remove_dir(TOP_DIR / mypy_report)
-        args.append('--html-report ' + mypy_report)
+        if report_dir is None:
+            remove_dir(TOP_DIR / mypy_report)
+            args.append('--xml-report ' + mypy_report)
+        else:
+            args.append('--xml-report ' + str(report_dir / 'mypy-coverage'))
     args.append(source_arg(src))
-    with c.cd(str(TOP_DIR)):
-        try:
-            c.run('mypy %s' % ' '.join(args), env=SRC_ENV, pty=True)
-        except UnexpectedExit as ex:
-            if ex.result.exited < 0:
-                print(ex)
+    out_path = None if report_dir is None else report_dir / 'mypy-log.txt'
+    out_stream = None if out_path is None \
+                      else open(out_path, 'w', encoding='utf-8')
+    try:
+        with c.cd(str(TOP_DIR)):
+            try:
+                c.run('mypy %s' % ' '.join(args),
+                      env=SRC_ENV, out_stream=out_stream, pty=True)
+            except UnexpectedExit as ex:
+                if ex.result.exited < 0:
+                    print(ex)
+    finally:
+        if out_stream is not None:
+            out_stream.close()
+    if results is not None:
+        errors = 0
+        with open(out_path, 'r', encoding='utf-8') as log:
+            for line in log:
+                if ' error: ' in line:
+                    errors += 1
+        with open(results, 'w', encoding='utf-8') as out:
+            out.write(f'result={"warning" if errors else "ok"}\n')
+            out.write(f'summary=mypy found {errors} errors\n')
+            out.write(f'report.{0 if errors else 2}={out_path}\n')
+            if report:
+                out.write(f'report.1={report_dir}/mypy-coverage\n')
 
 @task
 def isort(c, src=None):
